requirements.txt
absl-py==0.9.0
altgraph==0.17
appnope==0.1.0
astor==0.8.1
astunparse==1.6.3
asyncio==3.4.3
backcall==0.2.0
boto3==1.13.1
botocore==1.16.1
cachetools==4.1.0
certifi==2020.4.5.1
chardet==3.0.4
click==7.1.2
cx-Freeze==6.4.2
cycler==0.10.0
decorator==4.4.2
docopt==0.6.2
docutils==0.15.2
Flask==1.1.2
Flask-Cors==3.0.8
flaskwebgui==0.1.13
future==0.18.2
gast==0.2.2
gevent==20.4.0
google-auth==1.14.1
google-auth-oauthlib==0.4.1
google-pasta==0.2.0
greenlet==0.4.15
grpcio==1.28.1
gunicorn==20.0.4
h5py==2.10.0
idna==2.9
imageio==2.9.0
importlib-metadata==3.3.0
ipython==7.16.1
ipython-genutils==0.2.0
itsdangerous==1.1.0
jedi==0.17.2
Jinja2==2.11.2
jmespath==0.9.5
joblib==0.17.0
Keras==2.3.0
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
kiwisolver==1.2.0
lightgbm==3.0.0
macholib==1.14
Markdown==3.2.1
MarkupSafe==1.1.1
matplotlib==3.2.1
nltk==3.5
numpy==1.18.3
oauthlib==3.1.0
opencv-python==4.2.0.34
opt-einsum==3.2.1
pandas==1.0.3
parso==0.7.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==7.2.0
pipreqs==0.4.10
prompt-toolkit==3.0.5
protobuf==3.11.3
psutil==5.8.0
ptyprocess==0.6.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyCocoa==20.12.10
PyFladesk==1.1
Pygments==2.6.1
pyinstaller==4.1
pyinstaller-hooks-contrib==2020.10
pyparsing==2.4.7
PyQt5==5.15.2
PyQt5-sip==12.8.1
PyQtWebEngine==5.15.2
python-dateutil==2.8.1
python-resize-image==1.1.19
pytz==2020.1
pyvan==0.0.5
PyYAML==5.3.1
regex==2020.11.13
requests==2.23.0
requests-oauthlib==1.3.0
rsa==4.0
s3transfer==0.3.3
scikit-learn==0.23.2
scipy==1.4.1
six==1.14.0
tensorboard==2.1.1
tensorflow==2.1.0
tensorflow-estimator==2.1.0
termcolor==1.1.0
textblob==0.15.3
Theano==1.0.4
threadpoolctl==2.1.0
torch==1.6.0
torchvision==0.2.2.post3
tqdm==4.48.2
traitlets==4.3.3
typing-extensions==3.7.4.3
urllib3==1.25.9
wcwidth==0.2.5
webgl==0.0.1
Werkzeug==1.0.1
whichcraft==0.6.1
wrapt==1.12.1
yarg==0.1.9
zipp==3.4.0
